+++ sample-apps/react-native/ExpoTikTokApp/hooks/useCreateClient.ts
@@ -25,8 +25,7 @@ const tokenProviderFactory = (userId: string) => async () => {
 
 const CLIENT_OPTIONS = {
   configure_loggers_options: { default: { level: 'debug' } },
-  base_url: 'http://localhost:3030',
-  timeout: 10000
+  timeout: 10000,
 };
 
 export const useCreateClient = (user: LocalUser) => {